update Azure Storage queue WCF client
