+++ sdk/wcf/Microsoft.Azure.CoreWCF.Storage.Queues.Client/AzureQueueStorageClient.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureQueueStorageClient", "src\AzureQueueStorageClient.csproj", "{1CF1EE24-AF31-4F2D-9E3C-B588CD255DD3}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.WCF.AzureQueueStorage.Client", "src\Azure.WCF.AzureQueueStorage.Client.csproj", "{1CF1EE24-AF31-4F2D-9E3C-B588CD255DD3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AzureQueueStorageClient.Tests", "tests\AzureQueueStorageClient.Tests.csproj", "{5F5C3DC9-B558-4100-9192-1D2992A0C102}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.WCF.AzureQueueStorage.Client.Tests", "tests\Azure.WCF.AzureQueueStorage.Client.Tests.csproj", "{5F5C3DC9-B558-4100-9192-1D2992A0C102}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Azure.Storage.Queues", "..\..\storage\Azure.Storage.Queues\src\Azure.Storage.Queues.csproj", "{BA801FFE-4EDB-4389-BDFA-F62678432C7F}"
 EndProject
