+++ sdk/wcf/Microsoft.Azure.CoreWCF.Storage.Queues.Client/src/AzureQueueStorageChannelHelpers.cs
@@ -4,42 +4,9 @@
 using System;
 using System.Globalization;
 using System.ServiceModel;
-using System.ServiceModel.Channels;
 
-namespace Microsoft.ServiceModel.AQS
+namespace Azure.Storage.WCF.Channels
 {
-    /// <summary>
-    /// Collection of constants used by the AzureQueueStorage Channel classes
-    /// </summary>
-    internal static class AzureQueueStorageConstants
-    {
-        internal const string EventLogSourceName = "Microsoft.ServiceModel.AQS";
-        internal const string Scheme = "net.aqs";
-        private static MessageEncoderFactory s_messageEncoderFactory;
-        static AzureQueueStorageConstants()
-        {
-            s_messageEncoderFactory = new TextMessageEncodingBindingElement().CreateMessageEncoderFactory();
-        }
-
-        // ensure our advertised MessageVersion matches the version we're
-        // using to serialize/deserialize data to/from the wire
-        internal static MessageVersion MessageVersion
-        {
-            get
-            {
-                return s_messageEncoderFactory.MessageVersion;
-            }
-        }
-
-        internal static MessageEncoderFactory DefaultMessageEncoderFactory
-       {
-            get
-            {
-                return s_messageEncoderFactory;
-            }
-        }
-    }
-
     internal static class AzureQueueStorageChannelHelpers
     {
         /// <summary>
@@ -50,8 +17,8 @@ namespace Microsoft.ServiceModel.AQS
         internal static CommunicationException ConvertTransferException(Exception e)
         {
             return new CommunicationException(
-                string.Format(CultureInfo.CurrentCulture, 
-                "An error ({0}) occurred while transmitting message.", e.Message), 
+                string.Format(CultureInfo.CurrentCulture,
+                "An error ({0}) occurred while transmitting message.", e.Message),
                 e);
         }
