--- sdk/wcf/Microsoft.Azure.CoreWCF.Storage.Queues.Client/tests/TestService.cs
@@ -1,33 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-using System.ServiceModel;
-using System.Threading;
-
-namespace Microsoft.ServiceModel.AQS.Tests
-{
-    [ServiceContract]
-    public interface ITestContract
-    {
-        [OperationContract(IsOneWay = true)]
-        void Create(string name);
-    }
-
-    public class TestService : ITestContract
-    {
-        public TestService()
-        {
-            ManualResetEvent = new ManualResetEventSlim(false);
-        }
-
-        public void Create(string name)
-        {
-            if (string.IsNullOrEmpty(name))
-                throw new FaultException();
-
-            ManualResetEvent.Set();
-        }
-
-        public ManualResetEventSlim ManualResetEvent { get; }
-    }
-}
